tests/wpt/tests/html/canvas/offscreen/shadows/2d.shadow.outside.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.shadow.outside</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.shadow.outside</h1>
<p class="desc">Shadows of shapes outside the visible area can be offset onto the visible area</p>


<script>
var t = async_test("Shadows of shapes outside the visible area can be offset onto the visible area");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);
  ctx.shadowColor = '#0f0';
  ctx.shadowOffsetX = 100;
  ctx.fillRect(-100, 0, 25, 50);
  ctx.shadowOffsetX = -100;
  ctx.fillRect(175, 0, 25, 50);
  ctx.shadowOffsetX = 0;
  ctx.shadowOffsetY = 100;
  ctx.fillRect(25, -100, 50, 25);
  ctx.shadowOffsetY = -100;
  ctx.fillRect(25, 125, 50, 25);
  _assertPixel(canvas, 12,25, 0,255,0,255);
  _assertPixel(canvas, 87,25, 0,255,0,255);
  _assertPixel(canvas, 50,12, 0,255,0,255);
  _assertPixel(canvas, 50,37, 0,255,0,255);
  t.done();

});
</script>
